move make_output_wcs to resample, fix l3_wcs imports
