+++ romancal/resample/tests/test_resample.py
@@ -14,7 +14,7 @@ from romancal.assign_wcs.utils import add_s_region
 from romancal.datamodels import ModelLibrary
 from romancal.lib.tests.helpers import word_precision_check
 from romancal.resample import ResampleStep
-from romancal.resample._utils import make_output_wcs
+from romancal.resample.resample import make_output_wcs
 
 
 class WfiSca:
@@ -292,7 +292,7 @@ def test_resampledata_do_drizzle_many_to_one_default_no_rotation_single_exposure
     by checking that its extrema fall within the output WCS footprint.
 
     N.B.: since we are not providing the rotation parameter for the
-    resample_utils.make_output_wcs method, the output WCS will have
+    make_output_wcs method, the output WCS will have
     the same orientation (i.e. same PA) as the detector axes.
     """
 
@@ -322,7 +322,7 @@ def test_resampledata_do_drizzle_many_to_one_default_no_rotation_multiple_exposu
     by checking that its extrema fall within the output WCS footprint.
 
     N.B.: since we are not providing the rotation parameter for the
-    resample_utils.make_output_wcs method, the output WCS will have
+    make_output_wcs method, the output WCS will have
     the same orientation (i.e. same PA) as the detector axes.
     """
